+++ src/frontend/test-setup.ts
@@ -1,7 +1,7 @@
 import crypto from "@trust/webcrypto";
 import textEncoding = require("text-encoding");
 
-require("./src/test-e2e/mock-fetch");
+require("./src/test-e2e/setup-patch");
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
